Refactor specific UI components
